announcement: Move Sagir to SC Emeritus
I've been on a hiatus from the project for a while, and don't foresee
being able to come out of it soon-ish, owing to various personal
reasons. Till such a time as I'm unable to contribute actively, it's
best that I go emeritus and make space for new folks.

Thanks for having me aboard the rocketship! Ya'll have been a wonderful
team and a joy to work with. Sincere gratitude from the depth of my
heart to all the steering committee members, collaborators and
contributors! 🙏

Until next,
Here's to the next 50,000 stars! 🌟